moved sharedproviders -> commonproviders file
