--- postgis/describecollectionprovider.go
@@ -1,11 +1,0 @@
-package postgis
-
-import (
-	"net/http"
-	"oaf-server/codegen"
-	"oaf-server/provider"
-)
-
-func (pp *PostgisProvider) NewDescribeCollectionProvider(r *http.Request) (codegen.Provider, error) {
-	return provider.NewDescribeCollectionProvider(pp.Config)(r)
-}
